Unify styles for cafe page, add vertical centering of text
Addresses #11

Tweak styles for the text in each cafe__menu__group item on the cafe
page, decreasing heading size.

I have also unified the heading and body text into a single Divi
text element, to allow for easier styling. We set a minimum height on
the menu group, in order to easily calculate the top margin necessary
to vertically align the text for each cafe group.